# docs/swagger.yaml
basePath: /api-market-testnet
definitions:
...
  models.ConfirmationState:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - ConfirmationStateUnconfirmed
    - ConfirmationStateConfirmed
...
  request.BuildIdCoinsCommitRequest:
    properties:
      commitTxOutIndex:
        type: integer
      commitTxRaw:
        type: string
      orderId:
        type: string
    type: object
  request.BuildIdCoinsPreRequest:
    properties:
      amountPerMint:
        type: integer
      followersNum:
        type: integer
      issuerAddress:
        type: string
      issuerMetaId:
        type: string
      issuerSign:
        type: string
      liquidityPerMint:
        type: integer
      message:
        type: string
      networkFeeRate:
        type: integer
      tick:
        type: string
      ticker:
        type: string
    type: object
...
  request.IdCoinsMintCommitRequest:
    properties:
      commitTxOutInscribeIndex:
        type: integer
      commitTxOutMintIndex:
        type: integer
      commitTxRaw:
        type: string
      orderId:
        type: string
    type: object
  request.IdCoinsMintPreRequest:
    properties:
      networkFeeRate:
        type: integer
      outAddress:
        type: string
      outValue:
        type: integer
      tickId:
        type: string
    type: object
...
  respond.BuildIdCoinsCommitResp:
    properties:
      commitTxId:
        type: string
      orderId:
        type: string
      pinId:
        type: string
      revealTxId:
        type: string
      tickId:
        type: string
      txId:
        type: string
    type: object
  respond.BuildIdCoinsPreResp:
    properties:
      minerFee:
        type: integer
      orderId:
        type: string
      receiveAddress:
        type: string
      serviceFee:
        type: integer
      totalFee:
        type: integer
    type: object
...
  respond.FetchIdCoinsDeployCheckResp:
    properties:
      canDeploy:
        type: boolean
      msg:
        type: string
    type: object
  respond.FetchIdCoinsListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/respond.IdCoinsInfoResp'
        type: array
      total:
        type: integer
    type: object
  respond.FetchIdCoinsOpOrdersResp:
    properties:
      list:
        items:
          $ref: '#/definitions/respond.IdCoinsOpOrderInfoResp'
        type: array
      total:
        type: integer
    type: object
...
  respond.FetchOneIdCoinsMintOrderResp:
    properties:
      addressMintState:
        description: 0-unminted, 1-minted
        type: integer
      amtPerMint:
        type: string
      blockHeight:
        type: integer
      confirmationState:
        $ref: '#/definitions/models.ConfirmationState'
      decimals:
        type: string
      deployState:
        description: 0-pending, 1-success, 2-fail
        type: integer
      deployerAddress:
        type: string
      deployerMetaId:
        type: string
      deployerUserInfo:
        $ref: '#/definitions/respond.UserInfo'
      followersLimit:
        type: string
      liquidityPerMint:
        type: integer
      metaData:
        type: string
      mintCount:
        type: string
      mintState:
        description: 0-pending, 1-success, 2-fail
        type: integer
      opOrderType:
        type: string
      orderId:
        type: string
      payCheck:
        type: string
      pinCheck:
        type: string
      premineCount:
        type: string
      qual:
        type: string
      startBlockHeight:
        type: string
      tick:
        type: string
      tickId:
        type: string
      tickName:
        type: string
      timestamp:
        type: integer
      totalMinted:
        type: string
      txId:
        type: string
      usedPins:
        items:
          type: string
        type: array
    type: object
  respond.IdCoinsInfoResp:
    properties:
      amtPerMint:
        type: string
      blockHeight:
        type: string
      change24h:
        type: string
      decimals:
        type: string
      deployTime:
        type: integer
      deployerAddress:
        type: string
      deployerMetaId:
        type: string
      deployerUserInfo:
        $ref: '#/definitions/respond.UserInfo'
      floorPrice:
        type: string
      floorPriceUsd:
        type: string
      followersCount:
        type: integer
      followersLimit:
        type: string
      holders:
        type: integer
      isFollowing:
        type: boolean
      liquidityPerMint:
        type: integer
      marketCap:
        type: string
      marketCapUsd:
        type: string
      marketPrice:
        type: string
      marketPriceUsd:
        type: string
      metaData:
        type: string
      mintCount:
        type: string
      mintable:
        type: boolean
      mrc20Id:
        type: string
      ordersPool:
        type: integer
      ordersPrice:
        type: string
      payCheck: {}
      pinCheck: {}
      pinNumber:
        type: integer
      pool:
        type: integer
      premineCount:
        type: string
      price:
        type: string
      priceUsd:
        type: string
      qual: {}
      remaining:
        type: string
      supply:
        type: string
      tick:
        type: string
      tickId:
        type: string
      tokenName:
        type: string
      totalMinted:
        type: string
      totalSupply:
        type: string
      totalVolume:
        type: integer
      type:
        type: string
    type: object
  respond.IdCoinsMintCommitResp:
    properties:
      commitTxId:
        type: string
      orderId:
        type: string
      revealInscribeTxId:
        type: string
      revealMintTxId:
        type: string
    type: object
  respond.IdCoinsMintPreResp:
    properties:
      extra: {}
      orderId:
        type: string
      revealInscribeAddress:
        type: string
      revealInscribeFee:
        type: integer
      revealMintAddress:
        type: string
      revealMintFee:
        type: integer
      serviceAddress:
        type: string
      serviceFee:
        type: integer
      totalFee:
        type: integer
    type: object
  respond.IdCoinsOpOrderInfoResp:
    properties:
      amtPerMint:
        type: string
      blockHeight:
        type: integer
      confirmationState:
        $ref: '#/definitions/models.ConfirmationState'
      decimals:
        type: string
      deployState:
        description: 0-pending, 1-success, 2-fail
        type: integer
      deployerAddress:
        type: string
      deployerMetaId:
        type: string
      deployerUserInfo:
        $ref: '#/definitions/respond.UserInfo'
      followersLimit:
        type: string
      liquidityPerMint:
        type: integer
      metaData:
        type: string
      mintCount:
        type: string
      mintState:
        description: 0-pending, 1-success, 2-fail
        type: integer
      opOrderType:
        type: string
      orderId:
        type: string
      payCheck:
        type: string
      pinCheck:
        type: string
      premineCount:
        type: string
      qual:
        type: string
      startBlockHeight:
        type: string
      tick:
        type: string
      tickId:
        type: string
      tickName:
        type: string
      timestamp:
        type: integer
      totalMinted:
        type: string
      txId:
        type: string
      usedPins:
        items:
          type: string
        type: array
    type: object
...
  respond.UserInfo:
    properties:
      avatar:
        type: string
      name:
        type: string
    type: object
info:
  contact:
    name: API Support
  description: MetaID-Market API Service
  title: MetaID-Market API Service
  version: "1.0"
paths:
...
  /api/v1/id-coins/address/mint/order:
    get:
      description: Fetch address id-coins mint orders
      parameters:
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: tickId
        in: query
        name: tickId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.FetchOneIdCoinsMintOrderResp'
      summary: Fetch address id-coins mint orders
      tags:
      - IdCoins
  /api/v1/id-coins/coins-info:
    get:
      description: Fetch id-coins info
      parameters:
      - description: tickId
        in: query
        name: tickId
        required: true
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      - description: issuerAddress
        in: query
        name: issuerAddress
        type: string
      - description: address
        in: query
        name: address
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.IdCoinsInfoResp'
      summary: Fetch id-coins info
      tags:
      - IdCoins
  /api/v1/id-coins/coins-list:
    get:
      description: Fetch id-coins info list
      parameters:
      - description: address
        in: query
        name: address
        type: string
      - description: searchTick
        in: query
        name: searchTick
        type: string
      - description: 'supply/pool/timestamp, default:timestamp  '
        in: query
        name: orderBy
        type: string
      - description: default:-1
        in: query
        name: sortType
        type: integer
      - description: followerAddress
        in: query
        name: followerAddress
        type: string
      - description: default:0
        in: query
        name: cursor
        type: integer
      - description: default:10, max:50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.FetchIdCoinsListResp'
      summary: Fetch id-coins info list
      tags:
      - IdCoins
  /api/v1/id-coins/deploy/check/info:
    get:
      description: Fetch id-coins info
      parameters:
      - description: address
        in: query
        name: address
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.FetchIdCoinsDeployCheckResp'
      summary: Check id-coins deploy check info
      tags:
      - IdCoins
  /api/v1/id-coins/deploy/commit:
    post:
      description: Deploy id coins commit
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.BuildIdCoinsCommitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.BuildIdCoinsCommitResp'
      summary: Deploy id coins commit
      tags:
      - IdCoins
  /api/v1/id-coins/deploy/pre:
    post:
      description: Deploy id coins pre
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.BuildIdCoinsPreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.BuildIdCoinsPreResp'
      summary: Deploy id coins pre
      tags:
      - IdCoins
  /api/v1/id-coins/inscribe/orders:
    get:
      description: Fetch inscribe id-coins orders
      parameters:
      - description: default:deploy, mint
        in: query
        name: opOrderType
        required: true
        type: string
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: tickId
        in: query
        name: tickId
        type: string
      - description: default:0
        in: query
        name: cursor
        type: integer
      - description: default:10, max:50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.FetchIdCoinsOpOrdersResp'
      summary: Fetch inscribe id-coins orders
      tags:
      - IdCoins
  /api/v1/id-coins/mint/commit:
    post:
      description: Mint id coins commit
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.IdCoinsMintCommitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.IdCoinsMintCommitResp'
      summary: Mint id coins commit
      tags:
      - IdCoins
  /api/v1/id-coins/mint/pre:
    post:
      description: Mint id coins pre
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.IdCoinsMintPreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.IdCoinsMintPreResp'
      summary: Mint id coins pre
      tags:
      - IdCoins
...
schemes:
- https
swagger: "2.0"